server/pyproject.toml
[tool.poetry]
name = "homework_checker"
version = "0.1.0"
description = ""
authors = ["Michael Brauweiler <michael.brauweiler@posteo.de>"]

[tool.poetry.dependencies]
python = "^3.8, <3.9"
django = {extras = ["argon2"], version = "^3.1.2"}
python-environ = "^0.4.54"
djangorestframework = "^3.12.1"
markdown = "^3.2.2"
django-model-utils = "^4.0.0"
gunicorn = "^20.0.4"
dj-database-url = "^0.5.0"
pyyaml = "^5.3.1"
whitenoise = "^5.2.0"
python-dotenv = "^0.14.0"
pytest-cov = "^2.10.1"
pytest-asyncio = "^0.14.0"
pytest-django = "^3.10.0"
channels-redis = "^3.1.0"
channels = "^2.4.0"
django-autoslug = "^1.9.8"
psycopg2-binary = "^2.8.6"
django-cors-headers = "^3.5.0"
django-sendfile = "^0.3.11"
djangorestframework-simplejwt = "^4.5.0"

[tool.poetry.dev-dependencies]
isort = "^5.5.4"
black = "^20.8b1"
flake8 = "^3.8.3"
pytest = "^6.1.0"
pytest-django = "^3.10.0"
ipython = "^7.18.1"
factory_boy = "^3.1.0"

[tool.pytest.ini_options]
python_files = "test_*.py"
testpaths = ["tests"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"



